Added toggle of Emojis
Created a state to be able to modify if Emoji's (and potentially other things)
are on or off.

Add a function to toggle the state, using the ! convention that simply returns
the opposite value to an existing value.  So if the emoji is set to false, the !
character would return true.

Added a style for the button (well we are using text as a button in this case)
that we press to toggle the emojis.